face_service/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2

# ML / Vision
onnxruntime==1.19.2
opencv-python==4.10.0.84
numpy==1.26.4

# Vector Search
faiss-cpu==1.9.0
